+++ src/test/suite/commands/paredit.test.ts
@@ -1,8 +1,15 @@
 import assert from "assert";
-import { Position, Range, TextEditor } from "vscode";
+import { Position, Range, Selection, TextEditor } from "vscode";
 import { EmacsEmulator } from "../../../emulator";
 import { KillRing } from "../../../kill-yank/kill-ring";
-import { setEmptyCursors, setupWorkspace, assertCursorsEqual, assertTextEqual, clearTextEditor } from "../utils";
+import {
+  setEmptyCursors,
+  setupWorkspace,
+  assertCursorsEqual,
+  assertTextEqual,
+  clearTextEditor,
+  assertSelectionsEqual,
+} from "../utils";
 
 suite("paredit commands", () => {
   let activeTextEditor: TextEditor;
@@ -135,8 +142,8 @@ suite("paredit.kill-sexp", () => {
   test("killing inner parentheses with prefix argument", async () => {
     setEmptyCursors(activeTextEditor, [1, 0]);
 
-    emulator.universalArgument();
-    await emulator.universalArgumentDigit(2);
+    await emulator.universalArgument();
+    await emulator.subsequentArgumentDigit(2);
     await emulator.runCommand("paredit.killSexp");
 
     assertTextEqual(
@@ -160,6 +167,213 @@ suite("paredit.kill-sexp", () => {
     c d
   )`
     );
+  });
+});
+
+suite("paredit.backward-kill-sexp", () => {
+  const initialText = `(
+  (
+    a b
+  )
+  (
+    c d
+  )
+)`;
+  let activeTextEditor: TextEditor;
+  let emulator: EmacsEmulator;
+
+  setup(async () => {
+    activeTextEditor = await setupWorkspace(initialText);
+    const killRing = new KillRing(60);
+    emulator = new EmacsEmulator(activeTextEditor, killRing);
+  });
+
+  test("killing outer parentheses", async () => {
+    setEmptyCursors(activeTextEditor, [7, 1]);
+
+    await emulator.runCommand("paredit.backwardKillSexp");
+
+    assertTextEqual(activeTextEditor, "");
+    assertCursorsEqual(activeTextEditor, [0, 0]);
+
+    await clearTextEditor(activeTextEditor);
+
+    await emulator.runCommand("yank");
+
+    assertTextEqual(activeTextEditor, initialText);
+  });
+
+  test("killing inner parentheses continuously", async () => {
+    setEmptyCursors(activeTextEditor, [6, 3]);
+
+    await emulator.runCommand("paredit.backwardKillSexp");
+
+    assertTextEqual(
+      activeTextEditor,
+      `(
+  (
+    a b
+  )
+  
+)`
+    );
+    assertCursorsEqual(activeTextEditor, [4, 2]);
+
+    await emulator.runCommand("paredit.backwardKillSexp");
+
+    assertTextEqual(
+      activeTextEditor,
+      `(
+  
+)`
+    );
+    assertCursorsEqual(activeTextEditor, [1, 2]);
+
+    await clearTextEditor(activeTextEditor);
+
+    await emulator.runCommand("yank");
+
+    assertTextEqual(
+      activeTextEditor,
+      `(
+    a b
+  )
+  (
+    c d
+  )`
+    );
+  });
+
+  test("killing inner parentheses with prefix argument", async () => {
+    setEmptyCursors(activeTextEditor, [6, 3]);
+
+    await emulator.universalArgument();
+    await emulator.subsequentArgumentDigit(2);
+    await emulator.runCommand("paredit.backwardKillSexp");
+
+    assertTextEqual(
+      activeTextEditor,
+      `(
+  
+)`
+    );
+    assertCursorsEqual(activeTextEditor, [1, 2]);
+
+    await clearTextEditor(activeTextEditor);
+
+    await emulator.runCommand("yank");
+
+    assertTextEqual(
+      activeTextEditor,
+      `(
+    a b
+  )
+  (
+    c d
+  )`
+    );
+  });
+});
+
+suite("paredit.mark-sexp", () => {
+  const initialText = `(
+  (
+    a b
+  )
+  (
+    c d
+  )
+)`;
+  let activeTextEditor: TextEditor;
+  let emulator: EmacsEmulator;
+
+  setup(async () => {
+    activeTextEditor = await setupWorkspace(initialText);
+    const killRing = new KillRing(60);
+    emulator = new EmacsEmulator(activeTextEditor, killRing);
+  });
+
+  test("set mark at the outer parentheses", async () => {
+    setEmptyCursors(activeTextEditor, [0, 0]);
+
+    await emulator.runCommand("paredit.markSexp");
+
+    assertTextEqual(activeTextEditor, initialText);
+    assert.strictEqual(activeTextEditor.selections.length, 1);
+    assertSelectionsEqual(activeTextEditor, new Selection(0, 0, 7, 1));
+    assert.ok(emulator.isInMarkMode);
+
+    emulator.exitMarkMode();
+    activeTextEditor.selection = new Selection(0, 0, 0, 0);
+    emulator.popMark();
+    assertCursorsEqual(activeTextEditor, [7, 1]);
+  });
+
+  test("mark inner parentheses continuously", async () => {
+    setEmptyCursors(activeTextEditor, [1, 0]);
+    emulator.pushMark(activeTextEditor.selections.map((s) => s.active));
+
+    await emulator.runCommand("paredit.markSexp");
+
+    assertTextEqual(activeTextEditor, initialText);
+    assert.strictEqual(activeTextEditor.selections.length, 1);
+    assertSelectionsEqual(activeTextEditor, new Selection(1, 0, 3, 3));
+
+    await emulator.runCommand("paredit.markSexp");
+
+    assertTextEqual(activeTextEditor, initialText);
+    assert.strictEqual(activeTextEditor.selections.length, 1);
+    assertSelectionsEqual(activeTextEditor, new Selection(1, 0, 6, 3));
+
+    emulator.exitMarkMode();
+    activeTextEditor.selection = new Selection(0, 0, 0, 0);
+
+    emulator.popMark();
+    assertCursorsEqual(activeTextEditor, [6, 3]);
+    emulator.popMark();
+    assertCursorsEqual(activeTextEditor, [1, 0]);
+  });
+
+  test("mark inner parentheses with a positive prefix argument", async () => {
+    setEmptyCursors(activeTextEditor, [1, 0]);
+    emulator.pushMark(activeTextEditor.selections.map((s) => s.active));
+
+    await emulator.universalArgument();
+    await emulator.subsequentArgumentDigit(2);
+    await emulator.runCommand("paredit.markSexp");
+
+    assertTextEqual(activeTextEditor, initialText);
+    assert.strictEqual(activeTextEditor.selections.length, 1);
+    assertSelectionsEqual(activeTextEditor, new Selection(1, 0, 6, 3));
+
+    emulator.exitMarkMode();
+    activeTextEditor.selection = new Selection(0, 0, 0, 0);
+
+    emulator.popMark();
+    assertCursorsEqual(activeTextEditor, [6, 3]);
+    emulator.popMark();
+    assertCursorsEqual(activeTextEditor, [1, 0]);
+  });
+
+  test("mark inner parentheses with a negative prefix argument", async () => {
+    setEmptyCursors(activeTextEditor, [6, 3]);
+    emulator.pushMark(activeTextEditor.selections.map((s) => s.active));
+
+    await emulator.negativeArgument();
+    await emulator.subsequentArgumentDigit(2);
+    await emulator.runCommand("paredit.markSexp");
+
+    assertTextEqual(activeTextEditor, initialText);
+    assert.strictEqual(activeTextEditor.selections.length, 1);
+    assertSelectionsEqual(activeTextEditor, new Selection(6, 3, 1, 2));
+
+    emulator.exitMarkMode();
+    activeTextEditor.selection = new Selection(0, 0, 0, 0);
+
+    emulator.popMark();
+    assertCursorsEqual(activeTextEditor, [1, 2]);
+    emulator.popMark();
+    assertCursorsEqual(activeTextEditor, [6, 3]);
   });
 });
 
@@ -217,8 +431,8 @@ suite("paredit commands with prefix argument", () => {
   test("forwardSexp", async () => {
     setEmptyCursors(activeTextEditor, [0, 2]); // the right to `0`
 
-    emulator.universalArgument();
-    await emulator.universalArgumentDigit(2);
+    await emulator.universalArgument();
+    await emulator.subsequentArgumentDigit(2);
     await emulator.runCommand("paredit.forwardSexp");
 
     assert.strictEqual(activeTextEditor.selections.length, 1);
@@ -228,8 +442,8 @@ suite("paredit commands with prefix argument", () => {
   test("backwardSexp", async () => {
     setEmptyCursors(activeTextEditor, [0, 19]); // the left to `9`
 
-    emulator.universalArgument();
-    await emulator.universalArgumentDigit(2);
+    await emulator.universalArgument();
+    await emulator.subsequentArgumentDigit(2);
     await emulator.runCommand("paredit.backwardSexp");
 
     assert.strictEqual(activeTextEditor.selections.length, 1);